// src/core/products/dto/res/product-res.dto.ts
import { CategoryResDto } from 'src/core/categories/dto/res/category-res.dto'

export interface ProductResDto {
  id: number
  barcode: string | null
  name: string
  description: string | null
  retailPrice: number
  wholesalePrice: number
  wholesaleQty: number
  minStock: number
  image: string | null
  active: boolean
  category: CategoryResDto
}